LayoutTests/editing/deleting/merge-paragraph-from-p-with-style.html

<!DOCTYPE html>
<html>
<head>
<style type="text/css">
p {color: red;}
</style>
</head>
<body>
<p id="description">This tests deleting line break before p.
"helloworld" should both be in black text (no p element) and "world" should be italicized.</p>
<div id="test" contenteditable>
hello
<p style="font-style: italic;">world</p>
</div>
</div>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
<script>

function editingTest() {
    moveSelectionForwardByLineCommand();
    deleteCommand();

    Markup.description(document.getElementById('description').textContent);
    Markup.dump('test');
}

runEditingTest();
</script>
</body>
</html>
